Reflect.programelement interfaces' methods moved to LanguageElementHelper
